"sql reader class"

+++ src/MotorBikeSet.java
@@ -12,7 +12,7 @@ public class MotorBikeSet {
         return motorBikeSet;
     }
 
-    public void addMotorBike(int id, String name, int brandID, int year, int fuel, int price){
+    public void addMotorBike(int id, String name, int brandID, int year, int fuel, float price){
         MotorBike bike = new MotorBike(id, name, brandID, year, fuel, price);
         motorBikeSet.put(id, bike);
     }
